tests/wpt/web-platform-tests/html/canvas/element/fill-and-stroke-styles/2d.fillStyle.CSSHSL.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.fillStyle.CSSHSL</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.fillStyle.CSSHSL</h1>
<p class="desc">CSSHSL works as color input</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("CSSHSL works as color input");
_addTest(function(canvas, ctx) {

ctx.fillStyle = new CSSHSL(CSS.deg(180), 0.5, 0.5);
ctx.fillRect(0, 0, 100, 50);
_assertPixelApprox(canvas, 50,25, 64,191,191,255, "50,25", "64,191,191,255", 3);

const color = new CSSHSL(CSS.deg(180), 1, 1);
ctx.fillStyle = color;
_assertSame(ctx.fillStyle, '#ffffff', "ctx.fillStyle", "'#ffffff'");
ctx.fillRect(0, 0, 100, 50);
_assertPixel(canvas, 50,25, 255,255,255,255, "50,25", "255,255,255,255");
color.l = 0.5;
ctx.fillStyle = color;
_assertSame(ctx.fillStyle, '#00ffff', "ctx.fillStyle", "'#00ffff'");
ctx.fillRect(0, 0, 100, 50);
_assertPixel(canvas, 50,25, 0,255,255,255, "50,25", "0,255,255,255");

ctx.fillStyle = new CSSRGB(1, 0, 1).toHSL();
_assertSame(ctx.fillStyle, '#ff00ff', "ctx.fillStyle", "'#ff00ff'");
ctx.fillRect(0, 0, 100, 50);
_assertPixel(canvas, 50,25, 255,0,255,255, "50,25", "255,0,255,255");

color.h = CSS.deg(120);
color.s = 1;
color.l = 0.5;
ctx.fillStyle = color;
ctx.fillRect(0, 0, 100, 50);


});
</script>
